backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# AI/ML Libraries
openai==1.10.0
aisuite==0.1.2

# Research Tools
arxiv==2.1.0
wikipedia==1.4.0
tavily-python==0.3.0

# Web & Async
httpx==0.26.0
aiohttp==3.9.1
websockets==12.0

# Utils
python-dotenv==1.0.0
python-multipart==0.0.6
pyyaml==6.0.1

# Data Processing
pandas==2.1.4
numpy==1.26.3

# Export & Formatting
markdown==3.5.1
jinja2==3.1.3

# CORS & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing (optional)
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
